Prepared code modular by adding class-loader file
